+++ app/containers/DungeonScoresPage.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 
+import DarkOdinGrid from '../components/dungeonScores/DarkOdinGrid';
 import MagiciteGrid from '../components/dungeonScores/MagiciteGrid';
 import TormentGrid from '../components/dungeonScores/TormentGrid';
 import { Page } from './Page';
@@ -15,6 +16,9 @@ export default class DungeonScoresPage extends React.Component {
 
         <h4>Magicite</h4>
         <MagiciteGrid />
+
+        <h4>Dark Odin</h4>
+        <DarkOdinGrid />
 
         <div className={`alert alert-secondary ${styles.howToTip}`}>
           <strong>Tip:</strong> You may need to enter a dungeon in order to get complete time and